# docs/swagger.yaml
openapi: 3.0.0
info:
  title: ReactViteProKit API
  description: >
    Full-featured backend API built with Node.js, Express, MongoDB, Cloudinary.  
    Modular MVC design with Auth, User, Admin, Product, File, OTP modules.
    All responses follow a standard format with success status, data, and message fields.
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:5000/api
    description: 
  - url: https://example.com/api
    description: 
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token required for authenticated endpoints
paths:
...
  /files:
    post:
      tags:
        - File
      summary: Upload file
      description: >
        Uploads a file to Cloudinary cloud storage.
        Admin only endpoint.
        File is uploaded to 'uploads' folder on Cloudinary.
      operationId: uploadFile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: File to upload
      responses:
        "201":
          description: File uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      _id:
                        type: string
                        example: "507f1f77bcf86cd799439011"
                      filename:
                        type: string
                        description: Original filename
                        example: "document.pdf"
                      url:
                        type: string
                        format: uri
                        description: Cloudinary secure URL
                        example: "https://res.cloudinary.com/example/image.jpg"
                      publicId:
                        type: string
                        description: Cloudinary public ID for deletion
                        example: "uploads/abc123xyz"
                      uploadedBy:
                        type: string
                        description: User ID who uploaded the file
                        example: "507f1f77bcf86cd799439011"
                      createdAt:
                        type: string
                        format: date-time
                      updatedAt:
                        type: string
                        format: date-time
                  message:
                    type: string
                    example: "Operation successful"
        "400":
          description: No file provided
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Error message"
                  message:
                    type: string
                    example: "Error details"
        "401":
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Error message"
                  message:
                    type: string
                    example: "Error details"
        "403":
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Error message"
                  message:
                    type: string
                    example: "Error details"
  /files/{id}:
    get:
      tags:
        - File
      summary: Get file by ID
      description: >
        Retrieves file information by ID.
        Public endpoint - no authentication required.
      operationId: getFileById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: File ID (MongoDB ObjectId)
          example: "507f1f77bcf86cd799439011"
      responses:
        "200":
          description: File information
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      _id:
                        type: string
                        example: "507f1f77bcf86cd799439011"
                      filename:
                        type: string
                        description: Original filename
                        example: "document.pdf"
                      url:
                        type: string
                        format: uri
                        description: Cloudinary secure URL
                        example: "https://res.cloudinary.com/example/image.jpg"
                      publicId:
                        type: string
                        description: Cloudinary public ID for deletion
                        example: "uploads/abc123xyz"
                      uploadedBy:
                        type: string
                        description: User ID who uploaded the file
                        example: "507f1f77bcf86cd799439011"
                      createdAt:
                        type: string
                        format: date-time
                      updatedAt:
                        type: string
                        format: date-time
                  message:
                    type: string
                    example: "Operation successful"
        "404":
          description: File not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Error message"
                  message:
                    type: string
                    example: "Error details"

    delete:
      tags:
        - File
      summary: Delete file (admin only)
      description: >
        Deletes a file from Cloudinary and database.
        Admin only endpoint.
      operationId: deleteFileByAdmin
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: File ID (MongoDB ObjectId)
          example: "507f1f77bcf86cd799439011"
      responses:
        "200":
          description: File deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: "null"
                  message:
                    type: string
                    example: "Operation successful"
        "401":
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Error message"
                  message:
                    type: string
                    example: "Error details"
        "403":
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Error message"
                  message:
                    type: string
                    example: "Error details"
        "404":
          description: File not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Error message"
                  message:
                    type: string
                    example: "Error details"

 # /products:
  #   get:
  #     tags:
  #       - Product
  #     summary: Get all products
  #     description: >
  #       Retrieves a list of all products in the catalog.
  #       Public endpoint - no authentication required.
  #     operationId: getAllProducts
  #     responses:
  #       "200":
  #         description: List of products
  #         content:
  #           application/json:
  #             schema:
  #               allOf:
  #                 - $ref: "#/components/schemas/ApiResponse"
  #                 - type: object
  #                   properties:
  #                     data:
  #                       type: array
  #                       items:
  #                         $ref: "#/components/schemas/Product"

  #   post:
  #     tags:
  #       - Product
  #     summary: Create new product
  #     description: >
  #       Creates a new product.
  #       Admin only endpoint.
  #     operationId: addProduct
  #     security:
  #       - bearerAuth: []
  #     requestBody:
  #       required: true
  #       content:
  #         application/json:
  #           schema:
  #             type: object
  #             required:
  #               - name
  #               - price
  #             properties:
  #               name:
  #                 type: string
  #                 description: Product name
  #                 example: "Wireless Mouse"
  #               description:
  #                 type: string
  #                 description: Product description
  #                 example: "Ergonomic wireless mouse with 2.4GHz connection"
  #               price:
  #                 type: number
  #                 format: float
  #                 description: Product price
  #                 example: 29.99
  #               stock:
  #                 type: integer
  #                 description: Quantity in stock
  #                 example: 50
  #               imageUrl:
  #                 type: string
  #                 format: uri
  #                 description: Product image URL
  #                 example: "https://example.com/mouse.jpg"
  #     responses:
  #       "201":
  #         description: Product created successfully
  #         content:
  #           application/json:
  #             schema:
  #               allOf:
  #                 - $ref: "#/components/schemas/ApiResponse"
  #                 - type: object
  #                   properties:
  #                     data:
  #                       $ref: "#/components/schemas/Product"
  #       "400":
  #         description: Missing required fields (name and price)
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ErrorResponse"
  #       "401":
  #         description: Unauthorized - Invalid or missing token
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ErrorResponse"
  #       "403":
  #         description: Forbidden - Admin access required
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ErrorResponse"

  # /products/{id}:
  #   get:
  #     tags:
  #       - Product
  #     summary: Get product by ID
  #     description: >
  #       Retrieves a specific product by its ID.
  #       Public endpoint - no authentication required.
  #     operationId: getProductById
  #     parameters:
  #       - name: id
  #         in: path
  #         required: true
  #         schema:
  #           type: string
  #         description: Product ID (MongoDB ObjectId)
  #         example: "507f1f77bcf86cd799439011"
  #     responses:
  #       "200":
  #         description: Product details
  #         content:
  #           application/json:
  #             schema:
  #               allOf:
  #                 - $ref: "#/components/schemas/ApiResponse"
  #                 - type: object
  #                   properties:
  #                     data:
  #                       $ref: "#/components/schemas/Product"
  #       "404":
  #         description: Product not found
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ErrorResponse"

  #   put:
  #     tags:
  #       - Product
  #     summary: Update product
  #     description: >
  #       Updates a product's information.
  #       Admin only endpoint.
  #     operationId: editProduct
  #     security:
  #       - bearerAuth: []
  #     parameters:
  #       - name: id
  #         in: path
  #         required: true
  #         schema:
  #           type: string
  #         description: Product ID (MongoDB ObjectId)
  #         example: "507f1f77bcf86cd799439011"
  #     requestBody:
  #       required: true
  #       content:
  #         application/json:
  #           schema:
  #             type: object
  #             properties:
  #               name:
  #                 type: string
  #                 example: "Updated Product Name"
  #               description:
  #                 type: string
  #                 example: "Updated description"
  #               price:
  #                 type: number
  #                 format: float
  #                 example: 39.99
  #               stock:
  #                 type: integer
  #                 example: 100
  #               imageUrl:
  #                 type: string
  #                 format: uri
  #                 example: "https://example.com/newimage.jpg"
  #     responses:
  #       "200":
  #         description: Product updated successfully
  #         content:
  #           application/json:
  #             schema:
  #               allOf:
  #                 - $ref: "#/components/schemas/ApiResponse"
  #                 - type: object
  #                   properties:
  #                     data:
  #                       $ref: "#/components/schemas/Product"
  #       "401":
  #         description: Unauthorized - Invalid or missing token
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ErrorResponse"
  #       "403":
  #         description: Forbidden - Admin access required
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ErrorResponse"
  #       "404":
  #         description: Product not found
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ErrorResponse"

  #   delete:
  #     tags:
  #       - Product
  #     summary: Delete product
  #     description: >
  #       Deletes a product from the catalog.
  #       Admin only endpoint.
  #     operationId: deleteProduct
  #     security:
  #       - bearerAuth: []
  #     parameters:
  #       - name: id
  #         in: path
  #         required: true
  #         schema:
  #           type: string
  #         description: Product ID (MongoDB ObjectId)
  #         example: "507f1f77bcf86cd799439011"
  #     responses:
  #       "200":
  #         description: Product deleted successfully
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ApiResponse"
  #       "401":
  #         description: Unauthorized - Invalid or missing token
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ErrorResponse"
  #       "403":
  #         description: Forbidden - Admin access required
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ErrorResponse"
  #       "404":
  #         description: Product not found
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ErrorResponse"